requirements.txt
fastapi>=0.109.0
SQLAlchemy~=2.0.38
pydantic~=2.10.6
python-decouple~=3.4
psycopg2-binary>=2.9.9
setuptools>=68.0.0
uvicorn
alembic

